# apps/api/docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: Ownpace — managed API
  version: 0.1.0-rc.1
  summary: Migrate mail, calendar, contacts and files off US cloud, on your own schedule.
  description: |
    The multi-tenant managed edition. Every request is scoped to one tenant by
    the bearer token; the database enforces the same boundary again with
    row-level security, so a token for tenant A cannot read tenant B even if a
    route forgets to filter.

    **Two ideas run through the whole surface and explain most of its shape:**

    - *Nothing consequential happens without a person.* Deleting on the target,
      finishing a migration and moving billing money are gated behind explicit
      evidence, an owner/admin role, or both. Endpoints that could destroy data
      refuse by default and say what is missing.
    - *A failure is reported as itself.* Routes return the reason verbatim
      rather than an empty success. "I found nothing" and "I could not look"
      are different answers and never share a status code.

    Long-running work (sync, cutover, verification, applying a deletion) is
    enqueued, not performed in the request: those endpoints answer `202` with
    an identifier, and a matching read endpoint polls the outcome.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Ownpace
    url: https://github.com/Robbes/Ownpace
servers:
  - url: https://{host}
...
    variables:
      host:
        default: localhost:3001
...
security:
  - bearerAuth: []
...
paths:
...
  /api/setup/{side}/{provider}:
    get:
      tags: [Setup]
      summary: The provider setup checklist and where it has got to
      description: |
        What a person must do IN THE PROVIDER before a migration can read
        anything — create an app, grant scopes, get an administrator to
        authorise it, obtain a token. Steps are defined in code and keyed by
        `key`; state lives per tenant. A step nobody has answered has no stored
        row and reads as `open`, so a step added in a later release is open for
        everyone with no backfill.
      parameters:
        - { name: side, in: path, required: true, schema: { type: string, enum: [source, target] } }
        - { name: provider, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: The steps, their state, and the progress summary.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupChecklist' }
        '400': { description: 'The path was not /api/setup/{source|target}/{provider}.' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/setup/{side}/{provider}/{stepKey}:
    put:
      tags: [Setup]
      summary: Record what a person decided about one setup step
      description: |
        `done` and `skipped` are settled states and record who decided and
        when; setting a step back to `open` clears that, because an open step
        is by definition one nobody has answered. Answers the whole refreshed
        checklist, so the caller never has to re-read it.
      parameters:
        - { name: side, in: path, required: true, schema: { type: string, enum: [source, target] } }
        - { name: provider, in: path, required: true, schema: { type: string } }
        - { name: stepKey, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [state]
              properties:
                state: { type: string, enum: [open, done, skipped] }
      responses:
        '200':
          description: The refreshed checklist.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupChecklist' }
        '400': { description: 'The body was not a { state } object.' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: That step does not belong to this provider and side. }
...
  /api/decisions:
    get:
      tags: [Decisions]
      summary: Drift decisions awaiting an answer
      responses:
        '200':
          description: The decision queue.
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/decisions/{decisionId}/resolve:
    parameters:
      - $ref: '#/components/parameters/DecisionId'
    post:
      tags: [Decisions]
      summary: Answer a decision
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: Resolved.
          content:
            application/json:
              schema: { type: object }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/decisions/{decisionId}/dismiss:
    parameters:
      - $ref: '#/components/parameters/DecisionId'
    post:
      tags: [Decisions]
      summary: Set a decision aside
      description: |
        Set aside, not rejected. The distinction is kept because "I am not
        acting on this" and "this change was refused" are different claims
        about what happened.
      responses:
        '200':
          description: Dismissed.
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/decisions/presets:
    get:
      tags: [Decisions]
      summary: The standing answers
      description: |
        Per category, the answer applied automatically so the same question is
        not asked twice. A decision resolved this way says so, rather than
        appearing to have been considered by a person.
      responses:
        '200':
          description: The presets.
          content:
            application/json:
              schema: { type: object }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/decisions/presets/{category}:
    parameters:
      - name: category
        in: path
        required: true
        schema: { type: string }
    put:
      tags: [Decisions]
      summary: Set the standing answer for one category
      description: Owner/admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: Stored.
          content:
            application/json:
              schema: { type: object }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A tenant-scoped JWT. Local deployments sign it with `JWT_SECRET`;
        managed SSO validates against a JWKS instead.

        The token's subject must be a MEMBER of the tenant it names — an
        arbitrary subject with a valid signature is refused. Roles (`owner`,
        `admin`, `member`, `viewer`) gate the routes marked owner/admin above,
        for reads as well as writes on billing.
  parameters:
...
    DecisionId:
      name: decisionId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: The request was refused, with the reason.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, invalid, or non-member token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated, but this role may not do this.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: No such resource in this tenant.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    SetupChecklist:
      type: object
      properties:
        side: { type: string, enum: [source, target] }
        provider: { type: string }
        steps:
          type: array
          items:
            type: object
            properties:
              step:
                type: object
                properties:
                  key: { type: string }
                  titleKey: { type: string }
                  detailKey: { type: string }
                  yieldsKey: { type: string }
                  needsAnotherPerson: { type: boolean }
              state: { type: string, enum: [open, done, skipped] }
              decidedBy: { type: string }
              decidedAt: { type: string, format: date-time }
        progress:
          type: object
          properties:
            total: { type: integer }
            done: { type: integer }
            skipped: { type: integer }
            open: { type: integer }
            blockedOnOthers:
              type: integer
              description: Open steps needing an administrator — why a setup is stuck.
            complete: { type: boolean }
    Error:
      type: object
      description: Every refusal has this shape. `message` is meant to be shown to a person as-is.
      properties:
        error: { type: string, examples: ['Conflict'] }
        message:
          type: string
          examples: ['Mapping is paused — review the discovery counts and start it first (POST /start).']
      required: [error, message]
...